+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * `Style/GuardClause` doesn't report offenses in places where correction would make a line too long. ([@alexdowad][])
 * `Lint/DuplicateMethods` can find duplicate method definitions in many more circumstances, even across multiple files; however, it ignores definitions inside `if` or something which could be a DSL method. ([@alexdowad][])
 * A warning is printed if an invalid `EnforcedStyle` is configured. ([@alexdowad][])
+* [#1367](https://github.com/bbatsov/rubocop/issues/1367): New `Lint/IneffectiveAccessModifier` checks for access modifiers which are erroneously applied to a singleton method, where they have no effect. ([@alexdowad][])
 
 ### Bug Fixes